+++ src/api/Profile/apiPosts.js
@@ -1,0 +1,32 @@
+import {
+  addDoc,
+  arrayUnion,
+  collection,
+  doc,
+  updateDoc,
+} from 'firebase/firestore';
+import { firestore, storage } from '../../firebase/firebase';
+import { getDownloadURL, ref, uploadString } from 'firebase/storage';
+
+export const createEditPost = async (userId, caption, selectedFile, id) => {
+  const newPost = {
+    caption,
+    likes: [],
+    comments: [],
+    createdAt: Date.now(),
+    createdBy: userId,
+  };
+
+  // posts에 newPost추가, postRef 가져오기
+  const postDocRef = await addDoc(collection(firestore, 'posts'), newPost);
+  const userDocRef = doc(firestore, 'users', userId);
+  const imageRef = ref(storage, `posts/${postDocRef.id}`);
+
+  // 유저Ref에서 posts배열에 postDocRef.id추가
+  await updateDoc(userDocRef, { posts: arrayUnion(postDocRef.id) });
+
+  // postDocRef에 이미지URL추가
+  await uploadString(imageRef, selectedFile, 'data_url');
+  const downloadURL = await getDownloadURL(imageRef);
+  await updateDoc(postDocRef, { imageURL: downloadURL });
+};
